requirements.txt
Django>=4.2,<5.0
tensorflow>=2.12,<2.17
numpy>=1.24
Pillow>=9.5
matplotlib>=3.7
